# docs/swagger.yaml
basePath: /
definitions:
  dtos.CreateUserRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
      username:
        maxLength: 30
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
  dtos.LoginRequest:
    properties:
      password:
        minLength: 8
        type: string
      username:
        maxLength: 30
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
  dtos.VerifyOTPRequest:
    properties:
      otpkey:
        minLength: 1
        type: string
    required:
    - otpkey
    type: object
host: localhost:8080
info:
  contact:
    email: defabhishekkumarsingh@gmail.com
    name: API Support
  description: This is a production-grade REST API server built with Go.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go Resume API
  version: "1.0"
paths:
  /upload/:
    post:
      consumes:
      - multipart/form-data
      description: Upload a profile picture to Cloudinary
      parameters:
      - description: Profile Image
        in: formData
        name: profile_image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Upload Profile Image
      tags:
      - users
  /users/all:
    get:
      description: Get a paginated list of all users
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List All Users
      tags:
      - users
  /users/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user and receive a JWT and Refresh Token
      parameters:
      - description: Login Credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login User
      tags:
      - users
  /users/profile:
    get:
      description: Retrieve the profile of the currently authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get User Profile
      tags:
      - users
  /users/refreshaccesstoken:
    get:
      consumes:
      - application/json
      description: Generate a new JWT from a refresh token cookie
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Refresh Access Token
      tags:
      - users
  /users/register:
    post:
      consumes:
      - application/json
      description: Create a new user with username, email, and password
      parameters:
      - description: Registration Info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new User
      tags:
      - users
  /users/session/logout:
    post:
      description: Blacklist current JWT and revoke refresh token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Logout User
      tags:
      - users
  /users/verifyOtp:
    post:
      consumes:
      - application/json
      description: Verify the OTP submitted by the user
      parameters:
      - description: OTP Key
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.VerifyOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Verify OTP
      tags:
      - users
  /users/verifyemail:
    post:
      description: Send an OTP to the user's registered email
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Request Email OTP
      tags:
      - users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"